fix: move linkedin enrichment orchestration to the shell layer
autopilot_helpers/linkedin_enrich.rs (application layer) imported
tauri::AppHandle and called crate::commands::scrape::scrape_update_description
directly, violating the arch tests' R2 (no Tauri types below the shell
layer) and R7 (no upward layer imports) in tests/architecture.rs. Split
the module along the existing pure/orchestration boundary this crate
already uses elsewhere (rerank.rs, best_matches.rs):
select_linkedin_enrichment_targets, EnrichOutcome, and
classify_resolution stay pure at the application layer; the
AppHandle-touching fetch/rate-limit/write-back loop moves to a new
shell-layer submodule, commands/autopilot/linkedin_enrich.rs,
mirroring how commands/autopilot.rs already composes rerank/best_matches.
