+++ docs/knowledge/automation-domain.md
@@ -1,6 +1,6 @@
 # Automation domain (scraping + apply assistant + AI provider)
 
-Last updated: 2026-09-05 (task #16: backend-owned active AI provider store)
+Last updated: 2026-09-06 (task #16: backend-owned active AI provider store)
 
 Merged knowledge for `scraping-applier-expert` and `ai-provider-expert`. Source is authoritative for board/provider counts.
 
@@ -22,7 +22,7 @@ Merged knowledge for `scraping-applier-expert` and `ai-provider-expert`. Source 
 - **What the assistant does** — from a found or scraped job the user tailors a résumé + cover letter (`renderer/features/documents/components/TailorFlow/`, `renderer/store/generation-store/`), gets résumé-grounded application answers, opens the posting in the browser, and submits it themselves. Autopilot (`autopilot/`, `autopilot_scheduler`) **finds → ranks → notifies**; it never submits. Found-jobs PostingRow "Tailor" seeds the AI Generate workspace for any board.
 - **Autopilot board picker** — dynamically driven by the board catalog (`useBoardsCatalog()`), so new boards appear automatically without UI changes. The wizard board-select component (`StepTarget`) filters catalog entries by `listed` flag; includes the aggregator board.
 - **Autopilot scheduling** — **clock-anchored** (PR #266; was interval-based). `autopilot_scheduler.rs` computes the most-recent scheduled occurrence in local time (`chrono::Local`) via `last_occurrence_ms` / `is_due`; on launch it catches up to any missed run (single catch-up, no double-fire). Frequency modes: `daily` = HH:MM; `twice_daily` = HH:MM + 12 h; `hourly` = :MM; `manual` = no scheduling. Fields `scheduleHour` (0–23) and `scheduleMinute` (0–59) on the autopilot model — Zod-validated client-side, range-guarded in the store. Legacy records default to 09:00.
-- **Post-discovery LinkedIn description enrichment (issue #1114, LinkedIn-only slice)** — LinkedIn search results carry no description (`linkedin::api_client`'s comment on that field points here); a scheduled run's `autopilot_run` spawns `autopilot_helpers::linkedin_enrich::enrich_linkedin_descriptions` (fire-and-forget, via `tauri::async_runtime::spawn`) right after `record_run`, so it can never delay the run's own completion/notification. It re-fetches each still-blank LinkedIn found-job through the SAME single-URL resolver (`scraping::scrape_url::resolve`) the manual "re-check" and extension-import paths use, paced through `linkedin::rate_limiter::linkedin_rate_limiter()`, and writes any real description back via `commands::scrape::scrape_update_description` (the #1109 write-back mechanism) — never mutating anything else about the job. Selection is pure (`select_linkedin_enrichment_targets`, unit-tested): blank description + `board == "linkedin"`, not `is_new` (so a prior failed attempt retries), capped at `LINKEDIN_ENRICH_MAX` per run. A per-URL failure is logged and skipped, never surfaced as a run error. The other six description-blank boards named on `no_jd_text` (TheMuse, Comeet, Breezy, BambooHR, Pinpoint, Rippling) have no equivalent single-URL detail resolver yet — out of scope here.
+- **Post-discovery LinkedIn description enrichment (issue #1114, LinkedIn-only slice)** — LinkedIn search results carry no description (`linkedin::api_client`'s comment on that field points here); a scheduled run's `autopilot_run` spawns `commands::autopilot::linkedin_enrich::enrich_linkedin_descriptions` (L3 — Tauri/`AppHandle`-touching orchestration, fire-and-forget via `tauri::async_runtime::spawn`) right after `record_run`, so it can never delay the run's own completion/notification. The pure selection/decision logic it calls into stays one layer down in `autopilot_helpers::linkedin_enrich` (L2), per `docs/architecture-rules.md` R2/R7. It re-fetches each still-blank LinkedIn found-job through the SAME single-URL resolver (`scraping::scrape_url::resolve`) the manual "re-check" and extension-import paths use, paced through `linkedin::rate_limiter::linkedin_rate_limiter()`, and writes any real description back via `commands::scrape::scrape_update_description` (the #1109 write-back mechanism) — never mutating anything else about the job. Selection is pure (`select_linkedin_enrichment_targets`, unit-tested): blank description + `board == "linkedin"`, not `is_new` (so a prior failed attempt retries), capped at `LINKEDIN_ENRICH_MAX` per run. A per-URL failure is logged and skipped, never surfaced as a run error. The other six description-blank boards named on `no_jd_text` (TheMuse, Comeet, Breezy, BambooHR, Pinpoint, Rippling) have no equivalent single-URL detail resolver yet — out of scope here.
 - **Autopilot RUN filters** — a run applies three filters (date window, must-include keywords, min-match-score) that manual search does not; defaults realigned to manual parity via `AutopilotFilterSchema` Zod schema (packages/shared/src/schemas/index.ts) → generated Rust contract. Wizard `buildDefaults()` mirrors schema defaults. One-shot, marker-file-gated migration on first `autopilot_scheduler::start` calls `relax_legacy_filters_once` → `relax_legacy_filters` (pure logic) to loosen legacy saved autopilots per sentinel rules. Diagnostics: `scrape_diagnostics` surfaces per-board skip/error reasons into the autopilot run step log; `scrape_done` reports raw vs post-keyword-filter job counts.
 - **"Applied" tracking** — found-job enrichment: a found job is marked "applied" when a saved generation's `jobUrl` matches it (`commands/autopilot.rs: enrich_applied`). Separate paths now also auto-set application status: extension submit-watch form-tracking (opt-in, #687) and email-confirmation watching (#23, see section below). Each autopilot keeps an optional **base cover letter** (`coverLetter`) the assistant tailors per job.
 - **Assistant-notes provider resolution (task #16, behavior change)** — the per-record `assistant_provider/model/base_url` snapshot is gone; a scheduled/headless run now resolves the AI-reasoned-note provider via `Completer::from_active` — the **currently-active** provider at run time, not the one pinned when the schedule was created (owner-accepted semantics change; see [ADR-0012](decision-records/0012-ai-provider-base-url-provenance.md)). Missing/unknown/invalid provider → notes are skipped gracefully; the discovery run still completes. The renderer wizard's `assistantProvider`/`assistantModel`/`assistantBaseUrl` fields are now vestigial (always sent `undefined`).
